refactor: most members are in parent classes now
